# social-network-swagger-final.yaml
openapi: 3.0.0
info:
  title: musala soft
  version: 1.0.0
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    noauthAuth:
      type: http
      scheme: noauth
    bearerAuth:
      type: http
      scheme: bearer
security:
  - bearerAuth: []
...
paths:
  /api/v1/auth/authenticate:
    post:
      tags:
        - Auhthentication
      summary: Authenticate
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                email: deshan.sh@gmail.com
                password: '1234'
      security:
        - noauthAuth: []
      responses:
        '200':
          description: OK
          headers:
            Vary:
              schema:
                type: string
                example: Access-Control-Request-Headers
            X-Content-Type-Options:
              schema:
                type: string
                example: nosniff
            X-XSS-Protection:
              schema:
                type: integer
                example: '0'
            Cache-Control:
              schema:
                type: string
                example: no-cache, no-store, max-age=0, must-revalidate
            Pragma:
              schema:
                type: string
                example: no-cache
            Expires:
              schema:
                type: integer
                example: '0'
            X-Frame-Options:
              schema:
                type: string
                example: DENY
            Content-Type:
              schema:
                type: string
                example: application/json
            Transfer-Encoding:
              schema:
                type: string
                example: chunked
            Date:
              schema:
                type: string
                example: Sun, 26 May 2024 04:55:36 GMT
            Keep-Alive:
              schema:
                type: string
                example: timeout=60
            Connection:
              schema:
                type: string
                example: keep-alive
          content:
            application/json:
              schema:
                type: object
              example:
                body:
                  token: >-
                    eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJkZXNoYW5AZ21haWwuY29tIiwiaWF0IjoxNzE2Njk5MzM2LCJleHAiOjE3MTY3MDA3NzZ9.9nUlSSCjrHXtSx0IfaOfEK6J4q2vOJ8CppI5dZojHDA
                status: Success
                message: User Authenticated
  /api/v1/auth/register:
    post:
      tags:
        - Auhthentication
      summary: Registration
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                firstname: deshan
                lastname: salitha
                email: deshan.sh@gmail.com
                password: '1234'
      security:
        - noauthAuth: []
      responses:
        '200':
          description: OK
          headers:
            Vary:
              schema:
                type: string
                example: Access-Control-Request-Headers
            X-Content-Type-Options:
              schema:
                type: string
                example: nosniff
            X-XSS-Protection:
              schema:
                type: integer
                example: '0'
            Cache-Control:
              schema:
                type: string
                example: no-cache, no-store, max-age=0, must-revalidate
            Pragma:
              schema:
                type: string
                example: no-cache
            Expires:
              schema:
                type: integer
                example: '0'
            X-Frame-Options:
              schema:
                type: string
                example: DENY
            Content-Type:
              schema:
                type: string
                example: application/json
            Transfer-Encoding:
              schema:
                type: string
                example: chunked
            Date:
              schema:
                type: string
                example: Sun, 26 May 2024 04:56:25 GMT
            Keep-Alive:
              schema:
                type: string
                example: timeout=60
            Connection:
              schema:
                type: string
                example: keep-alive
          content:
            application/json:
              schema:
                type: object
              example:
                body:
                  token: >-
                    eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJkZXNoYW4uc2hAZ21haWwuY29tIiwiaWF0IjoxNzE2Njk5Mzg1LCJleHAiOjE3MTY3MDA4MjV9.rzkUqp2N73smSMkjGc9sa0yoaU_nPf9he8SMDwcm-IE
                status: Success
                message: User Authenticated
...